backend/requirements.txt
Django==4.2.19
djangorestframework==3.15.2
pillow==10.4.0
psycopg2-binary==2.9.10
pytest==8.3.4
pytest-django==4.10.0
python-decouple==3.8
reportlab==4.3.1